requirements.txt
# 📦 Загрузка переменных окружения из .env
python-dotenv>=1.0.0

# 🌐 Асинхронные HTTP-запросы (например, для погоды)
aiohttp>=3.9.0

# 🧪 Тестирование асинхронных хендлеров и FSM-сценариев
pytest>=7.0.0

# 🤖 Telegram-адаптер на базе aiogram 3.x
aiogram>=3.0.0